More subclasses of Client.ServerEvent
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@9233 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/database/variable/UserInterface.java
@@ -189,7 +189,7 @@ public interface UserInterface
      * Method to update the progress bar
      * @param pct the percentage done (from 0 to 100).
      */
-    public void setProgressValue(long pct);
+    public void setProgressValue(int pct);
 
     /**
      * Method to set a text message in the progress dialog.
